feat: Support semver in dependencies
